Derive initial interface selection based on policy
So far, any output module with a selection capability defaulted to
the first interface in the list as first pick. This uses the policy
configuration instead and thus allows to select which interface to
display first:

Examples:
 bmon -p 'em1,*'
 bmon -p 'eth*,lo,*'

The first rule that finds a matching interface is used and will
trigger initial selection.

Note that this is *ONLY* evaluated after the first read of the
statistics so if a more preferred interface appears later, the
selection will not be changed.

Signed-off-by: Thomas Graf <tgraf@suug.ch>